+++ packages/client/src/modals/instance/index.ts
@@ -1,6 +1,0 @@
-export * from './backendError';
-export * from './bundleReset';
-export * from './newFolder';
-export * from './selectPath';
-export * from './settingsReset';
-export * from './updateAvailable';