+++ packages/client/src/modals/model/index.ts
@@ -1,4 +1,0 @@
-export * from './createMetamodel';
-export * from './delete';
-export * from './download';
-export * from './edit';