+++ packages/client/src/stores/MainStore.tsx
@@ -10,9 +10,10 @@ import { makeAutoObservable, runInAction } from 'mobx';
 import { API } from '$api';
 import { defaultHotkeys } from '$data/hotkeys';
 import { parseHotkey } from '$hooks/useHotkey';
-import { InstanceBackendError } from '$modals/instance';
-import { ProjectUnsaved } from '$modals/project';
+import { InstanceBackendError } from '$modals/instance/backendError';
+import { ProjectUnsaved } from '$modals/project/unsaved';
 import { IS_ELECTRON } from '$utils/config';
+import { withoutTransitions } from '$utils/css';
 import { fuzzy, strIncludes } from '$utils/fuzzy';
 import { ConfigStore } from './ConfigStore';
 import { modalStore } from './ModalStore';
@@ -205,6 +206,10 @@ class MainStore {
       this.config.refresh(),
       setupStore.init(),
     ]);
+    withoutTransitions(() => {
+      document.documentElement.setAttribute('data-theme', this.theme);
+      document.documentElement.style.fontSize = `${this.config.data?.ui.fontSize || 16}px`;
+    });
     postMessage({ payload: 'removeLoading' }, '*');
     this.refreshModelCache();
     runInAction(() => {