+++ packages/client/src/views/project/index.tsx
@@ -19,3 +19,5 @@ export const Project: React.FC = observer(() => {
     </ProjectContext.Provider>
   );
 });
+
+export default Project;