+++ packages/client/src/views/settings/index.tsx
@@ -76,3 +76,5 @@ export const Settings: React.FC = observer(() => {
     </VarUI>
   );
 });
+
+export default Settings;